+++ fabric/src/main/java/net/frozenblock/wilderwild/data/WWRegistryProvider.java
@@ -77,8 +77,10 @@ final class WWRegistryProvider extends FabricDynamicRegistryProvider {
 		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.STRUCTURE_PROCESSOR_LIST_ADDITION)));
 		WWConstants.log("Adding finalized variant spawn injections to datagen", true);
 		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.VARIANT_SPAWN_INJECTION)));
-		WWConstants.log("Adding finalized variant biome environment attribute modifications to datagen", true);
+		WWConstants.log("Adding finalized biome environment attribute modifications to datagen", true);
 		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.BIOME_ENVIRONMENT_ATTRIBUTE_MODIFICATION)));
+		WWConstants.log("Adding finalized suffocation types to datagen", true);
+		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.SUFFOCATION_TYPE)));
 
 		// Wilder Wild Dynamic Registries
 		WWConstants.log("Adding finalized firefly colors to datagen", true);